* commit '655f75feeee4c396740a8e768888c7c8f0602cb8':
  Switch to a 4ms/1s timer

+++ Source/WebCore/platform/android/SharedTimerAndroid.cpp
@@ -51,7 +51,7 @@ void setSharedTimerFireTime(double fireTime)
 {
     long long timeInMs = static_cast<long long>((fireTime - WTF::currentTime()) * 1000);
 
-    ALOGV("setSharedTimerFireTime: in %ld millisec", timeInMs);
+    ALOGV("setSharedTimerFireTime: in %d millisec", (int)timeInMs);
     if (JavaSharedClient::GetTimerClient())
         JavaSharedClient::GetTimerClient()->setSharedTimer(timeInMs);
 }
